refactoring CLi
